Refactor project structure to move utility and model files
Moved `utils.rs` and `models.rs` into the `core` module
