+++ src/core/utils.rs
@@ -1,4 +1,4 @@
-use crate::FileState;
+
 use sha2::{Digest, Sha256};
 use std::fs;
 use std::io::{self, Write};
